New Feature/consolidated_backend/api/requirements.txt
# FastAPI & ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# MLflow client
mlflow>=2.9.0

# MinIO/S3
boto3>=1.33.0
minio>=7.2.0

# Redis (caching)
redis>=5.0.0

# Utils
python-dotenv>=1.0.0
httpx>=0.26.0

# CORS & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
